requirements.txt
python-telegram-bot==21.4
Flask==3.0.3
gunicorn==22.0.0